Cleanup and rename l1_handler to util
